requirements.txt
streamlit 
opencv-python 
openai 
faiss-cpu 
python-dotenv
moviepy